+++ Imperium/src/Interface/ImperiumUI/Windows/ObjectExplorer/ObjectExplorerWindow.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using Imperium.Interface.Common;
 using Imperium.Interface.ImperiumUI.Windows.ObjectExplorer.ObjectListEntry;
 using Imperium.Types;
+using Imperium.Util;
 using Imperium.Util.Binding;
 using TMPro;
 using UnityEngine;
@@ -66,7 +67,6 @@ internal class ObjectExplorerWindow : ImperiumWindow
 
     private float viewHeight;
     private float contentHeight;
-    private int entryCount;
     private const float entryHeight = 19;
     private float originalViewHeight;
 
@@ -75,6 +75,8 @@ internal class ObjectExplorerWindow : ImperiumWindow
 
     private List<ObjectCategory> categoryOrder;
     private Dictionary<ObjectCategory, CategoryDefinition> objectCategories;
+
+    private readonly ImpTimer periodicUpdateTimer = ImpTimer.ForInterval(1);
 
     private float previousScrollValue;
 
@@ -214,7 +216,7 @@ internal class ObjectExplorerWindow : ImperiumWindow
     private void InitEntryEngine()
     {
         originalViewHeight = Math.Abs(scrollRect.GetComponent<RectTransform>().sizeDelta.y);
-        entryCount = Mathf.CeilToInt(originalViewHeight / entryHeight) + 2;
+        var entryCount = Mathf.CeilToInt(originalViewHeight / entryHeight) + 2;
         viewHeight = entryHeight * entryCount;
 
         for (var i = 0; i < entryCount; i++)
@@ -307,17 +309,25 @@ internal class ObjectExplorerWindow : ImperiumWindow
             }
         }
 
-        playersCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.Players, 0).ToString()})";
-        entitiesCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.Entities, 0).ToString()})";
-        vehiclesCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.Vehicles, 0).ToString()})";
-        hazardsCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.Hazards, 0).ToString()})";
-        itemsCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.Items, 0).ToString()})";
-        ventsCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.Vents, 0).ToString()})";
-        vainShroudsCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.Vains, 0).ToString()})";
-        storyLogsCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.StoryLogs, 0).ToString()})";
-        othersCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.Other, 0).ToString()})";
-        outsideObjectsCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.OutsideObjects, 0).ToString()})";
+        playersCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.Players, 0)})";
+        entitiesCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.Entities, 0)})";
+        vehiclesCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.Vehicles, 0)})";
+        hazardsCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.Hazards, 0)})";
+        itemsCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.Items, 0)})";
+        ventsCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.Vents, 0)})";
+        vainShroudsCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.Vains, 0)})";
+        storyLogsCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.StoryLogs, 0)})";
+        othersCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.Other, 0)})";
+        outsideObjectsCount.text = $"({categoryCounts.GetValueOrDefault(ObjectCategory.OutsideObjects, 0)})";
     }
 
     private void OnScroll(Vector2 _) => StartCoroutine(refreshEntries(useCache: true));
+
+    private void Update()
+    {
+        if (periodicUpdateTimer.Tick())
+        {
+            Imperium.ObjectManager.RefreshLevelObjects();
+        }
+    }
 }